Remove support for Python 2.
This also adds Python 3.7 to the test list in tox.

This is backwards incompatible so it bumps the major version. Previously we had
been an unstable 0.x.y version, but this project has been live for a couple years
now and it's ready to go stable so we bump to 1.0.0.