Don't test the logging system.
After the switch to JSON logging, the code to test that logs were
set up and formatted correctly got really hairy. Deep into the weeds
of getting the right handlers set up for the tests, those test
modules were abandoned. It was kind of funky to reach all the way in
to test that in the first place, and it got ugly enough not to be
worth it.